Don't share the client fixture among tests.
When a client is used with on() or with_parent_id(), the parent_id value is set and remembered on the object.
This results in configured clients between tests, causing unexpected side effects.

This might however also be an issue in real life usage of the client, when multiple requests are performed using the same client instance. No issues of this type have ever been reported though.