Refactor properties that perform API requests into regular methods.
Most developers (me, at least) think of properties as being 'cheap'.
This is why I dislike the fact that we have object properties that actually
perform API requests, it sort of feels wrong to hide that amount of work
inside a reading a property. So now we have proper methods that perform
an API request to request data, and properties that just return (manipulated)
data from the existing object.

```
foobar = object.foobar
foobar = object.get_foobar()
```