Remove TODO on validating permissions by name: not a good idea.
After reviewing permissions, I noticed that many new permissions were added since the last time I checked.
If we add a list of valid permissions to the code, we need to release an update to the client every time
a new permission is added, and projects needing them are forced to upgrade to the new version without a
good reason.

The cons outweigh the pros. Syntax validation as we do now is good enough, I guess.