Implement a new iterator-only handler for result lists.
The current result list wrapping class `ObjectList` is intended for list handling,
but supports a mix of list- and  dictionary-related protocols: iteration, slices,
getitem (by key, not by position), etc.

One feature that you'd expect from a list that isn't supported, is getting quickly
from the start to the end of the list. You need to use a combination of `.has_next()`
and `.get_next()` to work through all (paginated) list items.

The old `Objectlist` supports key-based access to its members, but since the data is
paged, you can never be sure that the data is actually available in this page, or in
the next/previous one). Getting list length is also supported, but it only returns
the length of the current page, not a grand total. Both features are treacherous.

To improve on this, a new simple iterator-based approach was used, which transparently
fetches additional paginated data from the API. Things like `getitem` and `len()` are
no longer available, if you need these, they're easy to implement in custom code:

```python
    payments = client.payments.list()
    length = 0
    for item in payments:
        length += 1
        # Equivalent of payments["tr_12345"]
        if item.id == "tr_12345":
            break

    # Equivalent of len(payments)
    print(length)

    # An even better replacement for getting a specific payment would be:
    client.payments.get("tr_12345")
```

TODO:
- More tests
- Reversed() iterator, hoe does it work?
- Add logic to return an ResultListIterator where now an ObjectList is hardcoded
- Empty list results?