Remove incorrect type annotation from **kwargs
`**kwargs` should be annotated using only the type of the `value` in a `keyword=value` keyword argument.

Typing this as f.i. `**kwargs: str` means that all provided argument values need to be strings, which we can't support: `.list(limit=47)` uses an int, and it is valid usage. This leaves us with `Any` for now.