Resolve typing issues with get_resource_class()
This uncovers various issues with the implementation of `get_resource_class()`.
Now we have mypy passing and the code should work in theory.

But the only place where this method is used (in `ObjectList`), already shows
clearly that the current approach won't work: the generic way in which
`get_resource_class()` is currently used, doesn't allow for passing in random
required arguments.

We need to rethink the way ObjectList works, and hopefully get rid of
`get_resource_class()` entirely.